Save which lakes are train vs test to file
Separate out into a new function the part of get_train_test_data that
splits lakes into train and test sets. Call that function from the
script part of the file at the bottom. Then, lists of lakes that are in
the train and test sets are available to be saved.